% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{get_station_weather}
\alias{get_station_weather}
\title{Retrieve daily weather data from a specific station}
\usage{
get_station_weather(
  station_id,
  start_date = "2000-01-01",
  end_date = "3000-01-01"
)
}
\arguments{
\item{station_id}{station ID also known as WBANNO}

\item{start_date}{optional character (YYYY-MM-DD) or date parameter to filter for days beginning with start_date, inclusive}

\item{end_date}{optional character (YYYY-MM-DD) or date parameter to filter for days ending before start_date, inclusive}
}
\value{
a data frame with the following columns:
\itemize{
  \item{WBANNO}: {The station WBAN number}
  \item{state}: {the state location of the weather station}
  \item{station_name}: {the name of the station given by the state location vector}
  \item{LST_DATE}: {the local standard time date of the weather entry}
  \item{CRX_VN}: {the version number of the datalogger program at the entry time}
  \item{LONGITUDE}: {the longitude of the station using WGS-84}
  \item{LATITUDE}: {the latitude of the station using WGS-84}
  \item{T_DAILY_MAX}: {the maximum air temperature observed in C}
  \item{T_DAILY_MIN}: {the minimum air temperature observed in C}
  \item{T_DAILY_MEAN}: {the mean air temperature observed in C given by (T_DAILY_MAX + T_DAILY_MIN) / 2}
  \item{T_DAILY_AVG}: {the average air temperature observed in C}
  \item{P_DAILY_CALC}: {the total amount of precipitation in mm}
  \item{SOLARAD_DAILY}: {the total solar energy in MJ/meter^2 given by global radiation rates, conversion, and integration over time}
}
}
\description{
Get the National Centers for Environmental Information daily temperature and weather data for a
given weather station by its station ID with optional start and end date filters.
}
\examples{
# get daily temperature information for Ithaca, NY
ithaca_data_all <- get_station_weather(64758)
print(head(ithaca_data_all))
ithaca_data_early <- get_station_weather(64758, end_date = "2010-01-01")
print(head(ithaca_data_early))
ithaca_data_late <- get_station_weather(64758, start_date = "2020-01-01")
print(head(ithaca_data_late))
ithaca_data_specific <- get_station_weather(64758, start_date = "2010-01-01", end_date = "2020-01-01")
print(head(ithaca_data_specific))
}
