% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating.R
\name{get_yearly_cycle}
\alias{get_yearly_cycle}
\title{Estimate the yearly cycle for one station.
A yearly cycle is the expected temperature on each day of the year.}
\usage{
get_yearly_cycle(station_id, drop_leapdays = T)
}
\arguments{
\item{station_id}{station ID also known as WBANNO}

\item{drop_leapdays}{a logical evaluating to TRUE or FALSE indicating whether leapdays should be dropped or considered, respectively, in the returned dataframe.}
}
\value{
a data frame with the following columns:
\describe{
  \item{day_of_year}{The day of the year (1-365 or 366 if drop_leapdays is TRUE)}
  \item{avg_temp}{The expected average temperature on the day in C)}
}
}
\description{
The yearly cycle is taken from weather data for a
given weather station by its station ID
}
\examples{
# get yearly cycle for Ithaca, NY
ithaca_predictions <- get_yearly_cycle(64758)
print(head(ithaca_predictions))
ithaca_predictions_Leapyears <- get_yearly_cycle(64758,F)
print(head(ithaca_predictions_Leapyears))
}
