% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{interpolate_data}
\alias{interpolate_data}
\title{Interpolate values of X and Y via a gaussian processes model given a grid to interpolate to.}
\usage{
interpolate_data(datapoints, longitudes, latitudes, gridpoints, use_elev = F)
}
\arguments{
\item{longitudes}{a numeric of the longitudes associated with the points to interpolate.}

\item{latitudes}{a numeric of the latitudes associated with the points to interpolate.}

\item{gridpoints}{a series of sf points associated to the grid points to be interpolated with.}

\item{use_elev}{a logical of whether or not to calculate and use elevation for interpolations}

\item{toInterpolate}{a numeric of the datapoints which are to be interpolated.}
}
\value{
a dataframe containing the following columns:
\itemize{
  \item {interpolations}: {The numeric interpolated data points to plotted}
  \item {longitudes}: {A numeric of the longitudes to plot}
  \item {latitudes}: {A numeric of the latitudes to plot}
  \item {inUSA}: {A binary variable whether or not to plot (within contiguous USA)}
  }
}
\description{
Interpolate values of X and Y via a gaussian processes model given a grid to interpolate to.
}
\examples{
Interpolates a plot to the daily average temperature across the US
toInterpolate <- na.omit(daily_weather)
toInterpolate <- toInterpolate[!duplicated(toInterpolate$WBANNO),
                               c("LONGITUDE","LATITUDE","T_DAILY_AVG")]
interpolate_data(toInterpolate$T_DAILY_AVG,toInterpolate$LONGITUDE,toInterpolate$LATITUDE,
                 create_grid(resolution_X = 20,resolution_Y=20))
}
