% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_interpolations}
\alias{plot_interpolations}
\title{Plot points generated over a map of the contiguous US.}
\usage{
plot_interpolations(interpolated_data)
}
\arguments{
\item{interpolated_data}{the datapoints that are to be plotted with the following columns
\itemize{
  \item {interpolations}: {The interpolated data points to plotted}
  \item {longitudes}: {A numeric of the longitudes to plot}
  \item {latitudes}: {A numeric of the latitudes to plot}
  \item {inUSA}: {A binary variable whether or not to plot (within contiguous USA)}
  }}

\item{latitudes}{a numeric of the latitudes associated with the points to interpolate}
}
\value{
A plot with points overlaid.
}
\description{
Plot points generated over a map of the contiguous US.
}
\examples{
# get a plot of the USA with interpolated temperatures
point_map <- interpolate_data(toInterpolate$T_DAILY_AVG,toInterpolate$LONGITUDE,toInterpolate$LATITUDE,
                 create_grid(resolution_X = 20,resolution_Y=20))
plot_interpolations(point_map)
}
