% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station_info}
\alias{station_info}
\title{Weather station information}
\format{
a dataframe with 273 rows and 5 columns
\describe{
  \item {WBANNO}{"The station WBAN number"}
  \item {state}{the state location of the weather station}
  \item {station_name} {the name of the station given by the state location vector}
  \item {LST_DATE} {the local standard time date of the weather entry}
  \item {LONGITUDE} {the longitude of the station using WGS-84}
  \item {LATITUDE} {the latitude of the station using WGS-84}
}
}
\usage{
station_info
}
\description{
A dataset of National Centers for Environmental Information weather stations with name and location information recorded from 2020 to 2024
}
\keyword{datasets}
