% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating.R
\name{temperature_trend}
\alias{temperature_trend}
\title{Estimate the trend over time for annual temperatures.
This is done through fitting a model accounting for time elapsed in years and seasonality.}
\usage{
temperature_trend(station_id)
}
\arguments{
\item{station_id}{station ID also known as WBANNO}
}
\value{
A list of summary statistics for the fitted model.
}
\description{
Estimate the trend over time for annual temperatures.
This is done through fitting a model accounting for time elapsed in years and seasonality.
}
\examples{
# get yearly trend for Ithaca, NY
ithaca_model <- temperature_trend(64758)
print(ithaca_model)
}
